Refactor CRC calculations; Unsigned type for size and offset

+++ src/main/msp/msp_serial.h
@@ -92,8 +92,8 @@ typedef struct mspPort_s {
     uint16_t cmdMSP;
     uint8_t cmdFlags;
     mspVersion_e mspVersion;
-    int offset;
-    int dataSize;
+    uint_fast16_t offset;
+    uint_fast16_t dataSize;
     uint8_t checksum1;
     uint8_t checksum2;
 } mspPort_t;
